+++ packages/server/migrations/0017_drop_integration_connections.sql
@@ -1,0 +1,1 @@
+DROP TABLE "integration_connections" CASCADE;